Change button type, add form, add form controls
Changed the button to add a book to be a type button so it doesn't
submit data, and gave it an id so that we can easily target it in our
javascript.
Added a form with appropriate controls for the data we desire by users
who want to add a new book to the library, and gave it an id so that we
can differentiate it from the other button, this also doesn't submit
per say but will trigger javascript that will not only create a new
book but add it to the library.
I also gave the form an id so that I can style it and target it in the
js easily.
In addition to validation attributes and type attributes etc, I gave
specific name attribute values so that I can target the user input
correctly when creating a new book object.