Add id and for attributes, add class on form
To not only be able to easily target each input element using an id,
but also for accessibility concerns I've added id's to the input
elements and for attributes to their respective label elements with
the correct id value.
I also added a class to the form which will hide it by default, and
our javascript will toggle this class when the Add book button is
clicked.