Change modal position, style accordingly
I added some comments for sectioning purposes.
I positioned the modal so that when it appears it's about centered on
the viewport, and gave it a default class that hides it so the
javascript will toggle this class on and off the element when the add
book button is clicked.
I also gave the modal parent element a z-index value so it'll overlap
the preexisting books in the collection and allow the user to focus on
adding  a new book.
I styled the fieldset within the modal's form to make it a little more
user friendly, and positioned its children items using grid-area so we
have a strict placement.