Add an event listener to showModal to display it
This event listener when clicked toggles a class which allows us to
show the form element to the user, and even though that means if they
press the button again that it hides the modal, the modal may be
covering the button. So I'll need to create a button in the modal to
close the modal now.