Adjust modal grid for new close modal button
I added an extra row so that the button could sit on its own row, and
used a justify-self of end to move it further to the right than just
using the second column for it.
I also applied a background color of red to hopefully communicate
that this button has a negative side effect, which is it closes the
element although it does not wipe user input that is currently residing
in the controls.