Add functionality to add new book to library
This function does not add the new book into the array as the array
will be wiped upon a reload, that functionality would need a database
or localStorage, and I will add that later.
So for now what I just added will pull values from the modal and build
a book object and add that object to a card and that card into our
books container so visually they are adding a book to this library.
Tomorrow I will complete the last 2 steps of this project which is
adding a remove button to each book card, and a change read status
button so users can edit books that they've read since adding the
books to their library.