Remove extra function invocations, arguments, etc
While working on adding data-book-number I realized that previous code
was very incorrect, books that were being added by users were actually
calling the book constructor multiple times, once within the event
listener and again inside addBookToLibrary, and the value being passed
to addBookToLibrary was very incorrect as well. So I went ahead and
fixed that behavior. In the submitNewBook eventListener we now call
addBookToLibrary passing it the correct values from the user input
fields.
I also added delete buttons to each book card that is made via the
javascript and had to create a new global variable, which I know is
bad, which keeps track of how many books have been added into the
library, this value is used to keep track of elements created and
relating them to the books in the array. To get this to work as I
wanted it to I had to change the for of loop in displayMyLibraryBooks
to a regular for loop which uses an iterator variable so we can pass
that integer into the buildBookCard function so we can set up our
data-attribute with the appropriate number. On a similar note when a
user adds a book via the add book button we call buildBookCard and pass
it the last object in the array and the value of the totalBooks variable
counter -1 because indexing starts at 0 in javascript.
So each time a book is added to the library the totalBooks variable
value is incremented by 1.