Add grid to book elements, deleteButton justify
Added grid with explicit columns for the book elements our js
dynamically creates to give it a more strict structure which should
help if there is a lot of info in any of the fields, although I could
just limit user input lengths to keep this under control.
Positioned the delete button at the end of its row in the book element
container as I wanted it offset to a side so a user has to look away
from the flow of the books content, which might require more thought
before they click it.