Omit totalBooks, target specific object
Omitted totalBooks as it was saving the value at time of assignment not
referencing the new value of myLibrary.length... oops.
I added a specific index with the array that is being passed to the
buildBookCard within the submitNewBook button listener, the function
was not working as intended because it was receiving an array of 1
object, when it needed to specifically receive 1 object.