Change data-book-number
After testing with deleting existing books and then creating new books
and looking at their data-book-number values I've realized that by
using index values we are creating a problem, once we delete a book
and then add a new book that new book might have the same
data-book-number value as another book.
If we create unique numbers that can be added as a property into the
book objects and not rely on the array index values then we shouldn't
have any problems.
And sadly I'm going to use what I used earlier, a counter that
increments each time a book is created, and since deleting a bookCard
element does indeed cause a shift/repaint then we don't need to re
render the entire group of bookCards as I had feared.
Changed the delete buttons functionality to pull the identifier out and
then use findIndex to get the array index of this specific book and
then use splice to remove it from the array.
I also removed an unnecessary use of bookCount in the  submitNewBook
listener, it was not needed at all.