Replace Book constructor function with class
Classes are easier to read as long as you know how they work, and setting up methods on the prototype are easier to code than with constructors.
The other difference was that classes aren't hoisted in the same way that constructor functions are so we had to move the first two calls to the Book class to after the class declaration.