Add comments, styling
Added a comment to explain how the listener is working.
Added styling for input fields when they are not valid and when they are valid.
I'm seeing "An invalid form control with name"..." is not focusable" error for the book title, author, and pages input fields, and I'm unsure why.